[OSD-8942] add ocm agent e2e test
